Reintroduced getHelper to WebTestCase
